+++ tests/unit/ssh_keys/conftest.py
@@ -9,7 +9,8 @@ def ssh_key_response():
             "name": "My ssh key",
             "fingerprint": "b7:2f:30:a0:2f:6c:58:6c:21:04:58:61:ba:06:3b:2f",
             "public_key": "ssh-rsa AAAjjk76kgf...Xt",
-            "labels": {}
+            "labels": {},
+            "created": "2016-01-30T23:50:00+00:00"
         }
     }
 
@@ -23,14 +24,16 @@ def two_ssh_keys_response():
                 "name": "SSH-Key",
                 "fingerprint": "b7:2f:30:a0:2f:6c:58:6c:21:04:58:61:ba:06:3b:2f",
                 "public_key": "ssh-rsa AAAjjk76kgf...Xt",
-                "labels": {}
+                "labels": {},
+                "created": "2016-01-30T23:50:00+00:00"
             },
             {
                 "id": 2324,
                 "name": "SSH-Key",
                 "fingerprint": "b7:2f:30:a0:2f:6c:58:6c:21:04:58:61:ba:06:3b:2f",
                 "public_key": "ssh-rsa AAAjjk76kgf...Xt",
-                "labels": {}
+                "labels": {},
+                "created": "2016-01-30T23:50:00+00:00"
             }
         ]
     }
@@ -59,6 +62,7 @@ def response_update_ssh_key():
             "name": "New name",
             "fingerprint": "b7:2f:30:a0:2f:6c:58:6c:21:04:58:61:ba:06:3b:2f",
             "public_key": "ssh-rsa AAAjjk76kgf...Xt",
-            "labels": {}
+            "labels": {},
+            "created": "2016-01-30T23:50:00+00:00"
         }
     }